requirements.txt
certifi==2019.11.28
Django==2.1.5
django-registration-redux==2.2
olefile==0.46
Pillow==5.4.1
pytz==2019.3
django-haystack==2.8.1
Whoosh==2.7.4
